Hotel.java Upload
Currently has two known bugs. The first one is ab error on line 464, "Cannot Find Symbol". The second one is more of a logic issue. Haven't found a way to keep track of "currentUserID" so it will give syntax errors on compile. Other functions seem to work fine so far.